requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.1
langchain-google-genai==0.0.5
llama-index==0.9.8
llama-index-embeddings-huggingface==0.1.4
tenacity==8.2.3
pydantic==2.5.3
google-generativeai==0.3.1
requests==2.31.0
tqdm==4.66.1
numpy==1.24.3
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0
langchain-community==0.0.13
python-docx==1.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
